# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoRa KT API",
    "version": "1.0.0",
    "description": "REST API for team management, checkpoints, RFID access control, LoRa telemetry, and map data."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative to application root"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "webhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Optional server-level secret for the ingest endpoint. When LORA_WEBHOOK_SECRET is set (and not the default placeholder), every POST /api/ingest must include this header."
      }
    },
    "schemas": {
...
      "GroupCheckpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "position"
        ]
      },
      "CheckpointGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupCheckpoint"
            }
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/groups/order": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Update group order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "group_ids"
                ],
                "properties": {
                  "group_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Ordered list of all group IDs."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated"
          }
        }
      }
    },
...
    "/api/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "List checkpoint groups",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groups": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CheckpointGroup"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Create group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "description": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "checkpoint_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created group"
          }
        }
      }
    },
    "/api/groups/{group_id}": {
      "parameters": [
        {
          "name": "group_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Group detail",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointGroup"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "groups"
        ],
        "summary": "Update group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "description": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "checkpoint_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated group"
          }
        }
      },
      "delete": {
        "tags": [
          "groups"
        ],
        "summary": "Delete group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "put": {
        "tags": [
          "groups"
        ],
        "summary": "Replace group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "description": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "checkpoint_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced group"
          }
        }
      }
    },
...
  }
}